# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ShopOS Auto Repair Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /customers:
    get:
      operationId: getCustomers
      tags: [customers]
      summary: List customers
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of customers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerListResponse"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Create a customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomerInput"
      responses:
        "201":
          description: Customer created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Get customer by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    put:
      operationId: updateCustomer
      tags: [customers]
      summary: Update customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomerInput"
      responses:
        "200":
          description: Customer updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    delete:
      operationId: deleteCustomer
      tags: [customers]
      summary: Delete customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /customers/{id}/vehicles:
    get:
      operationId: getCustomerVehicles
      tags: [customers]
      summary: Get vehicles for a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer vehicles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Vehicle"
  /customers/{id}/statement:
    get:
      operationId: getCustomerStatement
      tags: [customers]
      summary: Get billing statement for customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer billing statement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerStatement"

  # ---- VEHICLES ----
...
  /appointments:
    get:
      operationId: getAppointments
      tags: [appointments]
      summary: List appointments
      parameters:
        - name: date
          in: query
          schema:
            type: string
            format: date
        - name: status
          in: query
          schema:
            type: string
            enum: [scheduled, confirmed, in_progress, completed, cancelled, no_show]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: List of appointments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppointmentListResponse"
    post:
      operationId: createAppointment
      tags: [appointments]
      summary: Create an appointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAppointmentInput"
      responses:
        "201":
          description: Appointment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
  /appointments/{id}:
    get:
      operationId: getAppointment
      tags: [appointments]
      summary: Get appointment by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Appointment details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
    put:
      operationId: updateAppointment
      tags: [appointments]
      summary: Update appointment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAppointmentInput"
      responses:
        "200":
          description: Appointment updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
    delete:
      operationId: deleteAppointment
      tags: [appointments]
      summary: Delete appointment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ---- PAYMENTS ----
...
  /employees:
    get:
      operationId: getEmployees
      tags: [employees]
      summary: List employees
      parameters:
        - name: role
          in: query
          schema:
            type: string
        - name: active
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of employees
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Employee"
    post:
      operationId: createEmployee
      tags: [employees]
      summary: Create employee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEmployeeInput"
      responses:
        "201":
          description: Employee created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
  /employees/{id}:
    get:
      operationId: getEmployee
      tags: [employees]
      summary: Get employee by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Employee details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
    put:
      operationId: updateEmployee
      tags: [employees]
      summary: Update employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEmployeeInput"
      responses:
        "200":
          description: Employee updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
    delete:
      operationId: deleteEmployee
      tags: [employees]
      summary: Delete employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /employees/{id}/clock-in:
    post:
      operationId: clockIn
      tags: [time-entries]
      summary: Clock in employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Clocked in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimeEntry"
  /employees/{id}/clock-out:
    post:
      operationId: clockOut
      tags: [time-entries]
      summary: Clock out employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Clocked out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimeEntry"

  # ---- TIME ENTRIES ----
...
components:
  schemas:
...
    Customer:
      type: object
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        city:
          type: string
        state:
          type: string
        zip:
          type: string
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        vehicleCount:
          type: integer
        totalBilled:
          type: number
        totalPaid:
          type: number
      required:
        - id
        - firstName
        - lastName
        - createdAt
        - updatedAt
    CreateCustomerInput:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        city:
          type: string
        state:
          type: string
        zip:
          type: string
        notes:
          type: string
      required:
        - firstName
        - lastName
    CustomerListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Customer"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - data
        - total
        - page
        - limit
    CustomerStatement:
      type: object
      properties:
        customer:
          $ref: "#/components/schemas/Customer"
        invoices:
          type: array
          items:
            $ref: "#/components/schemas/Invoice"
        totalBilled:
          type: number
        totalPaid:
          type: number
        balance:
          type: number
      required:
        - customer
        - invoices
        - totalBilled
        - totalPaid
        - balance
    Vehicle:
      type: object
      properties:
        id:
          type: integer
        customerId:
          type: integer
        vin:
          type: string
        licensePlate:
          type: string
        year:
          type: integer
        make:
          type: string
        model:
          type: string
        trim:
          type: string
        color:
          type: string
        mileage:
          type: integer
        engineType:
          type: string
        transmissionType:
          type: string
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        customer:
          $ref: "#/components/schemas/Customer"
      required:
        - id
        - customerId
        - year
        - make
        - model
        - createdAt
        - updatedAt
...
    LineItem:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [labor, part, fee, discount]
        description:
          type: string
        quantity:
          type: number
        unitPrice:
          type: number
        total:
          type: number
        partNumber:
          type: string
        inventoryItemId:
          type: integer
      required:
        - type
        - description
        - quantity
        - unitPrice
        - total
...
    Invoice:
      type: object
      properties:
        id:
          type: integer
        invoiceNumber:
          type: string
        customerId:
          type: integer
        vehicleId:
          type: integer
        repairOrderId:
          type: integer
        estimateId:
          type: integer
        status:
          type: string
          enum: [draft, sent, paid, overdue, void]
        notes:
          type: string
        subtotal:
          type: number
        taxRate:
          type: number
        taxAmount:
          type: number
        discountAmount:
          type: number
        total:
          type: number
        amountPaid:
          type: number
        balance:
          type: number
        dueDate:
          type: string
          format: date
        lineItems:
          type: array
          items:
            $ref: "#/components/schemas/LineItem"
        payments:
          type: array
          items:
            $ref: "#/components/schemas/Payment"
        customer:
          $ref: "#/components/schemas/Customer"
        vehicle:
          $ref: "#/components/schemas/Vehicle"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - invoiceNumber
        - customerId
        - status
        - subtotal
        - taxRate
        - taxAmount
        - discountAmount
        - total
        - amountPaid
        - balance
        - createdAt
        - updatedAt
...
    Appointment:
      type: object
      properties:
        id:
          type: integer
        customerId:
          type: integer
        vehicleId:
          type: integer
        assignedToId:
          type: integer
        status:
          type: string
          enum: [scheduled, confirmed, in_progress, completed, cancelled, no_show]
        serviceType:
          type: string
        description:
          type: string
        scheduledAt:
          type: string
          format: date-time
        estimatedDuration:
          type: integer
        notes:
          type: string
        customer:
          $ref: "#/components/schemas/Customer"
        vehicle:
          $ref: "#/components/schemas/Vehicle"
        assignedTo:
          $ref: "#/components/schemas/Employee"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - customerId
        - status
        - serviceType
        - scheduledAt
        - createdAt
        - updatedAt
    CreateAppointmentInput:
      type: object
      properties:
        customerId:
          type: integer
        vehicleId:
          type: integer
        assignedToId:
          type: integer
        status:
          type: string
          enum: [scheduled, confirmed, in_progress, completed, cancelled, no_show]
        serviceType:
          type: string
        description:
          type: string
        scheduledAt:
          type: string
          format: date-time
        estimatedDuration:
          type: integer
        notes:
          type: string
      required:
        - customerId
        - serviceType
        - scheduledAt
    AppointmentListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Appointment"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - data
        - total
        - page
        - limit
    Payment:
      type: object
      properties:
        id:
          type: integer
        invoiceId:
          type: integer
        amount:
          type: number
        method:
          type: string
          enum: [cash, check, credit_card, debit_card, paypal, square, stripe, other]
        referenceNumber:
          type: string
        notes:
          type: string
        paidAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - invoiceId
        - amount
        - method
        - paidAt
        - createdAt
...
    Employee:
      type: object
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [technician, service_advisor, manager, admin, receptionist]
        hourlyRate:
          type: number
        active:
          type: boolean
        hireDate:
          type: string
          format: date
        notes:
          type: string
        clockedIn:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - firstName
        - lastName
        - role
        - active
        - createdAt
        - updatedAt
    CreateEmployeeInput:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [technician, service_advisor, manager, admin, receptionist]
        hourlyRate:
          type: number
        active:
          type: boolean
        hireDate:
          type: string
          format: date
        notes:
          type: string
      required:
        - firstName
        - lastName
        - role
    TimeEntry:
      type: object
      properties:
        id:
          type: integer
        employeeId:
          type: integer
        clockIn:
          type: string
          format: date-time
        clockOut:
          type: string
          format: date-time
        totalHours:
          type: number
        notes:
          type: string
        employee:
          $ref: "#/components/schemas/Employee"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - employeeId
        - clockIn
        - createdAt
...